+++ test/Ord.ts
@@ -5,10 +5,11 @@ import { fold } from '../src/Monoid'
 import { pipe } from '../src/function'
 import * as B from '../src/boolean'
 import * as S from '../src/string'
+import * as N from '../src/number'
 
 describe('Ord', () => {
   it('getTupleOrd', () => {
-    const O = _.getTupleOrd(S.Ord, _.ordNumber, B.Ord)
+    const O = _.getTupleOrd(S.Ord, N.Ord, B.Ord)
     assert.deepStrictEqual(O.compare(['a', 1, true], ['b', 2, true]), -1)
     assert.deepStrictEqual(O.compare(['a', 1, true], ['a', 2, true]), -1)
     assert.deepStrictEqual(O.compare(['a', 1, true], ['a', 1, false]), 1)
@@ -24,7 +25,7 @@ describe('Ord', () => {
     ]
     const M = _.getMonoid<T>()
     const sortByFst = pipe(
-      _.ordNumber,
+      N.Ord,
       _.contramap((x: T) => x[0])
     )
     const sortBySnd = pipe(
@@ -49,14 +50,8 @@ describe('Ord', () => {
     ])
   })
 
-  it('ordNumber', () => {
-    assert.deepStrictEqual(_.ordNumber.compare(1, 2), -1)
-    assert.deepStrictEqual(_.ordNumber.compare(2, 1), 1)
-    assert.deepStrictEqual(_.ordNumber.compare(2, 2), 0)
-  })
-
   it('clamp', () => {
-    const clampNumber = _.clamp(_.ordNumber)
+    const clampNumber = _.clamp(N.Ord)
     assert.deepStrictEqual(clampNumber(1, 10)(2), 2)
     assert.deepStrictEqual(clampNumber(1, 10)(10), 10)
     assert.deepStrictEqual(clampNumber(1, 10)(20), 10)
@@ -65,7 +60,7 @@ describe('Ord', () => {
   })
 
   it('between', () => {
-    const betweenNumber = _.between(_.ordNumber)
+    const betweenNumber = _.between(N.Ord)
     assert.deepStrictEqual(betweenNumber(1, 10)(2), true)
     assert.deepStrictEqual(betweenNumber(1, 10)(10), true)
     assert.deepStrictEqual(betweenNumber(1, 10)(20), false)
@@ -74,7 +69,7 @@ describe('Ord', () => {
   })
 
   it('getDualOrd', () => {
-    const O = _.getDualOrd(_.ordNumber)
+    const O = _.getDualOrd(N.Ord)
     assert.deepStrictEqual(O.compare(1, 2), 1)
     assert.deepStrictEqual(O.compare(2, 1), -1)
     assert.deepStrictEqual(O.compare(2, 2), 0)
@@ -87,19 +82,19 @@ describe('Ord', () => {
   })
 
   it('leq', () => {
-    assert.deepStrictEqual(_.leq(_.ordNumber)(0, 1), true)
-    assert.deepStrictEqual(_.leq(_.ordNumber)(1, 1), true)
-    assert.deepStrictEqual(_.leq(_.ordNumber)(2, 1), false)
+    assert.deepStrictEqual(_.leq(N.Ord)(0, 1), true)
+    assert.deepStrictEqual(_.leq(N.Ord)(1, 1), true)
+    assert.deepStrictEqual(_.leq(N.Ord)(2, 1), false)
   })
 
   it('geq', () => {
-    assert.deepStrictEqual(_.geq(_.ordNumber)(0, 1), false)
-    assert.deepStrictEqual(_.geq(_.ordNumber)(1, 1), true)
-    assert.deepStrictEqual(_.geq(_.ordNumber)(2, 1), true)
+    assert.deepStrictEqual(_.geq(N.Ord)(0, 1), false)
+    assert.deepStrictEqual(_.geq(N.Ord)(1, 1), true)
+    assert.deepStrictEqual(_.geq(N.Ord)(2, 1), true)
   })
 
   it('fromCompare', () => {
-    const O1 = _.fromCompare(_.ordNumber.compare)
+    const O1 = _.fromCompare(N.Ord.compare)
     assert.deepStrictEqual(O1.equals(0, 1), false)
     assert.deepStrictEqual(O1.equals(1, 1), true)
     interface A {
@@ -108,7 +103,7 @@ describe('Ord', () => {
     let nbCall = 0
     const O2 = _.fromCompare<A>((a, b) => {
       nbCall += 1
-      return _.ordNumber.compare(a.x, b.x)
+      return N.Ord.compare(a.x, b.x)
     })
     const a1 = { x: 1 }
     const a2 = { x: 1 }